+++ srcs/ft_strncmp.c
@@ -6,15 +6,15 @@
 /*   By: alejhern <alejhern@student.42barcelon      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/12 20:22:53 by alejhern          #+#    #+#             */
-/*   Updated: 2024/06/12 22:18:26 by alejhern         ###   ########.fr       */
+/*   Updated: 2024/06/15 20:45:40 by alejhern         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strncmp(char *s1, char *s2, unsigned int n)
+int	ft_strncmp(char *s1, char *s2, size_t n)
 {
-	unsigned int	index;
+	size_t	index;
 
 	i = 0;
 	if (n == 0)
